+++ CHANGELOG.md
@@ -13,6 +13,7 @@ _2022-06-01_
   toString and an `.or()` method for fallbak/default values
 - feat(ts): Alias `toString` as `getName` on all basic value printers — to
   hack a nicer DX when using es-in-css values with styled-components.
+- fix: Check for existstence of global.require before clearing cache
 
 ## 0.4.0